[Refactor]: 코드 컨벤션 준수하여 패키지 분리, 패키지명 수정
