+++ presentation/src/main/kotlin/ac/dnd/bookkeeping/android/presentation/common/state/ApplicationState.kt
